migrate to modern initcontainer based sidecars (#112)
Kuberentes 1.29 adds support for sidecar containers that can be ensured to be started/running before the main container is up _and_ be stopped after.

See https://kubernetes.io/docs/concepts/workloads/pods/sidecar-containers/#sidecar-containers-and-pod-lifecycle

This is helpful to ensure that our tls cert/termination is available through the life of a pod and may help to ensure that datadog stats near the very begginning and very end of a pod (first/last 10s) have a chance to be captured/flushed respectively.